requirements.txt
absl-py==2.0.0
accelerate==0.23.0
aiohttp==3.8.5
aiosignal==1.3.1
anyio==4.2.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
bert-score==0.3.13
bitsandbytes==0.41.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.1
cycler==0.11.0
datasets==2.14.5
deepspeed==0.10.3
dill==0.3.7
distro==1.9.0
docker-pycreds==0.4.0
docutils==0.20.1
einops==0.6.1
evaluate==0.4.0
exceptiongroup==1.2.0
filelock==3.12.4
fire==0.5.0
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.37
h11==0.14.0
hjson==3.1.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.17.2
idna==3.4
importlib-resources==6.1.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.1
ninja==1.11.1
nltk==3.8.1
numpy==1.26.0
openai==1.8.0
packaging==23.1
pandas==2.1.1
pathtools==0.1.2
Pillow==10.0.1
protobuf==4.24.3
psutil==5.9.5
py-cpuinfo==9.0.0
pyarrow==13.0.0
pydantic==1.10.12
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
randomname==0.2.1
regex==2023.8.8
requests==2.31.0
responses==0.18.0
rouge-score==0.1.2
safetensors==0.4.0
scipy==1.11.2
sentencepiece==0.1.99
sentry-sdk==1.31.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
statistics==1.0.3.5
sympy==1.12
termcolor==2.3.0
tiktoken==0.5.2
tokenizers==0.13.3
torch==2.0.0
tqdm==4.66.1
transformers==4.33.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
wandb==0.15.11
wordcloud==1.9.3
xxhash==3.3.0
yarl==1.9.2
zipp==3.17.0
